requirements/local.txt
## Import Base Requirements
-r _base.txt

## Local Specific Requirements
django-debug-toolbar==1.9.1
flake8==3.5.0
